pyproject.toml
[tool.poetry]
name = "brb"
version = "0.1.0"
description = "Django web app to store your terminal logs"
authors = ["Lee Sutton <leesutton1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pytz = "=2018.9"
python-slugify = "=2.0.1"
Pillow = "=5.4.1"
argon2-cffi = "=19.1.0"
whitenoise = "=4.1.2"
redis = "*"
kombu= "*"
celery = "=4.2.1"
flower = "=0.9."
django = "=2.0.13"
django-environ = "=0.4.5"
django-model-utils = "=3.1.2"
django-allauth = "=0.38.0"
django-crispy-forms = "=1.7.2"
django-redis = "=4.10.0"
djangorestframework = "=3.9.1"
coreapi = "=2.3.3"
Werkzeug = "=0.14.1"
ipdb = "=0.11"
Sphinx = "=1.8.4"
psycopg2 = "=2.7.4"
gunicorn = "=19.9.0"
raven = "=6.10."
django-storages = "^1.7"
django-anymail = "^6.0"

[tool.poetry.dev-dependencies]
mypy = "=0.670"
pytest = "=4.2.0"
flake8 = "=3.7.5"
coverage = "=4.5."
factory-boy = "=2.11.1"
django-debug-toolbar = "=1.11"
django-extensions = "=2.1.5"
django-coverage-plugin = "=1.6.0"
pytest-django = "=3.4.7"
pytest-watch = "^4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

